New location provider
Replaced the old and confusing location provider with a new implementation, this fixed the provider dependancy issues where the app was trying to look for restaurants with coordinates that had not been given yet.

The new location provider works with a hardcoded set of location data that it passes to the restaurantsprovider. Tested that the location changing is handled smoothly by adding a couple of buttons to shift location between Helsinki and Kuopio